Adjust dates in database population
This changes behaviour in cases where there has been days without
changes to the data set and then a callsign expires or changes
state. Then we adjust the date to the day before the new changeset if
the previous data set is older than that. Thus, no gaps are left to
the history of events.